requirements.txt
feedparser>=6.0.11
requests>=2.31.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
python-dateutil>=2.9.0
PyYAML>=6.0.1
pytest>=8.0.0
pytest-mock>=3.12.0
responses>=0.25.0
